refactor : CoordinateRangeDto를 dto.request 패키지 내부로 이동
