+++ src/main/java/day6/fullbang/service/PlaceService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import day6.fullbang.domain.Place;
-import day6.fullbang.dto.CoordinateRangeDto;
+import day6.fullbang.dto.request.CoordinateRangeDto;
 import day6.fullbang.repository.PlaceRepository;
 import lombok.RequiredArgsConstructor;
